backend/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
sqlalchemy==1.4.49
asyncpg==0.28.0
alembic==1.10.3
psycopg2-binary==2.9.7
pydantic==1.10.7
httpx==0.24.1
celery[redis]==5.3.1
redis==4.5.3
celery-redbeat==2.1.1
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.24.3
cryptography==41.0.2
aiofiles==23.1.0
pytz==2023.3
elasticsearch==8.11.0

# AI/ML dependencies for semantic search
sentence-transformers==2.2.2
huggingface-hub==0.14.1
faiss-cpu==1.7.4
huggingface-hub==0.14.1

# NLP
spacy==3.7.2
nltk==3.8.1

# Clustering for phase detection
# Note: hdbscan disabled due to wheel build issues in this image. PhaseDetector
# falls back to scikit-learn clustering when HDBSCAN isn't available.

pytest==7.4.0
